csharp: Change HighLevelCallback length to int
Because Array.copy will only take an int, not a long.

+++ csharp/IPConnection.cs
@@ -1349,7 +1349,7 @@ namespace Tinkerforge
 	internal class HighLevelCallback
 	{
 		internal object data = null;
-		internal long length = 0;
+		internal int length = 0;
 	}
 
 	internal class BrickDaemon : Device
